Added CLI, added logging, changes to match in rest
Added a basic command line interface to input file paths and models.
Added a struct that contains the tree and sequence data and implemented appropriate
checks for the given information, namely that sequence IDs match tree tip IDs.
Included logging into the code to show meaningful messages about the progress
and the errors.
Edited error reporting to work with the logging scheme.